updated globus_compute
