+++ workflows/globus_compute/0_remote_adder.py
@@ -1,12 +1,20 @@
 from globus_compute_sdk import Executor
 
+# This script is intended to be run from your remote machine
+
+# Scripts adapted from Globus Compute docs
+# https://globus-compute.readthedocs.io/en/latest/quickstart.html
+
 # First, define the function ...
 def add_func(a, b):
     return a + b
 
-tutorial_endpoint_id = 'c0396551-2870-45f2-a2aa-70991eb120a4'
+# Paste your endpoint id here, e.g.
+# endpoint_id = 'c0396551-2870-45f2-a2aa-70991eb120a4'
+endpoint_id = ''
+
 # ... then create the executor, ...
-with Executor(endpoint_id=tutorial_endpoint_id) as gce:
+with Executor(endpoint_id=endpoint_id) as gce:
     # ... then submit for execution, ...
     future = gce.submit(add_func, 5, 10)
